reorganize dashboard component, adding some state switching logic
